Removing unused options and better setting of defaults.
Default for -S SlicingFactor:Slice should be 1:0, so that  pre-selected
unpaired kmers can be read in without chance of getting the slicing
wrong. (Slicing can still be used if input kmers haven't been selected
down.)
Likewise for -P "positions" should be "*", so that pre-selected SNPmer
pairs can be read without messing up. (Position can still be specified
when on-the-fly selection is preferred.)